moving things around to simplify namespaces and folder layout. adding more docs.
