+++ src/base/BenchmarkResult.php
@@ -2,8 +2,6 @@
 
 namespace markhuot\craftpest\base;
 
-use Illuminate\Support\Collection;
-use yii\db\Command;
 use craft\debug\Module;
 
 /**
@@ -21,9 +19,9 @@ use craft\debug\Module;
  *   ->assertNoDuplicateQueries();
  * ```
  * 
- * @see \markhuot\craftpest\traits\Benchmark
+ * @see \markhuot\craftpest\test\Benchmark
  */
-class Benchmark
+class BenchmarkResult
 {
     protected $manifestCache;
