+++ src/dom/Form.php
@@ -3,7 +3,7 @@
 namespace markhuot\craftpest\dom;
 
 use markhuot\craftpest\http\RequestBuilder;
-use markhuot\craftpest\traits\Dd;
+use markhuot\craftpest\test\Dd;
 use markhuot\craftpest\web\TestableResponse;
 use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\DomCrawler\Field\ChoiceFormField;
