+++ src/test/LocalVolumes.php
@@ -2,13 +2,22 @@
 
 namespace markhuot\craftpest\test;
 
-use craft\events\ModelEvent;
-use craft\helpers\ProjectConfig;
-use Symfony\Component\Process\Process;
-use yii\base\Application;
-use yii\base\Event;
-use yii\db\Transaction;
-
+/**
+ * # Local Volumes
+ *
+ * If you're using an older version of Craf that does not support swappable filesystems you can
+ * use the `LocalVolumes` trait to convert any S3 volumes in to local folder volumes during
+ * test.
+ *
+ * Add the `LocalVolumes` trait to your `Pest.php`'s `uses()` call like so,
+ *
+ * ```php
+ * uses(
+ *   markhuot\craftpest\test\TestCase::class,
+ *   markhuot\craftpest\test\LocalVolumes::class,
+ * );
+ * ```
+ */
 trait LocalVolumes {
 
     function setUpLocalVolumes() {
