+++ stubs/init/Pest.php
@@ -13,7 +13,7 @@
 
 uses(
     markhuot\craftpest\test\TestCase::class,
-    markhuot\craftpest\test\RefreshesDatabase::class,
+    \markhuot\craftpest\test\RefreshesDatabase::class,
 )->in('./');
 
 /*
